moved more related functionality from shipcurse.c to flyingship.c
