+++ src/flyingship.h
@@ -11,8 +11,12 @@
 extern "C" {
 #endif
 
+    void register_flyingship(void);
+
     int sp_flying_ship(castorder * co);
 
+    struct curse *shipcurse_flyingship(struct ship *sh, struct unit *mage,
+        double power, int duration);
     int levitate_ship(struct ship *sh, struct unit *mage, double power,
         int duration);
     bool flying_ship(const ship * sh);
