+++ src/spells/shipcurse.c
@@ -78,41 +78,6 @@ static struct curse_type ct_stormwind = { "stormwind",
 CURSETYP_NORM, 0, NO_MERGE, cinfo_ship
 };
 
-static int flyingship_read(storage * store, curse * c, void *target)
-{
-    ship *sh = (ship *)target;
-    c->data.v = sh;
-    if (global.data_version < FOSS_VERSION) {
-        sh->flags |= SF_FLYING;
-        return 0;
-    }
-    assert(sh->flags & SF_FLYING);
-    return 0;
-}
-
-static int flyingship_write(storage * store, const curse * c,
-    const void *target)
-{
-    const ship *sh = (const ship *)target;
-    assert(sh->flags & SF_FLYING);
-    return 0;
-}
-
-static int flyingship_age(curse * c)
-{
-    ship *sh = (ship *)c->data.v;
-    if (sh && c->duration == 1) {
-        freset(sh, SF_FLYING);
-        return 1;
-    }
-    return 0;
-}
-
-static struct curse_type ct_flyingship = { "flyingship",
-CURSETYP_NORM, 0, NO_MERGE, cinfo_ship, NULL, flyingship_read,
-flyingship_write, NULL, flyingship_age
-};
-
 static struct curse_type ct_nodrift = { "nodrift",
 CURSETYP_NORM, 0, (M_DURATION | M_VIGOUR), cinfo_shipnodrift
 };
@@ -121,35 +86,10 @@ static struct curse_type ct_shipspeedup = { "shipspeedup",
 CURSETYP_NORM, 0, 0, cinfo_ship
 };
 
-curse *shipcurse_flyingship(ship * sh, unit * mage, double power, int duration)
-{
-    static const curse_type *ct_flyingship = NULL;
-    if (!ct_flyingship) {
-        ct_flyingship = ct_find("flyingship");
-        assert(ct_flyingship);
-    }
-    if (curse_active(get_curse(sh->attribs, ct_flyingship))) {
-        return NULL;
-    }
-    else if (is_cursed(sh->attribs, C_SHIP_SPEEDUP, 0)) {
-        return NULL;
-    }
-    else {
-        /* mit C_SHIP_NODRIFT haben wir kein Problem */
-        curse *c =
-            create_curse(mage, &sh->attribs, ct_flyingship, power, duration, 0.0, 0);
-        c->data.v = sh;
-        if (c && c->duration > 0) {
-            sh->flags |= SF_FLYING;
-        }
-        return c;
-    }
-}
 
 void register_shipcurse(void)
 {
     ct_register(&ct_stormwind);
-    ct_register(&ct_flyingship);
     ct_register(&ct_nodrift);
     ct_register(&ct_shipspeedup);
 }
